Add columns and restructure files
