+++ app-modules/ai/routes/api.php
@@ -34,9 +34,6 @@
 </COPYRIGHT>
 */
 
-use AdvisingApp\Ai\Http\Controllers\QnaAdvisorAuthenticationConfirmController;
-use AdvisingApp\Ai\Http\Controllers\QnaAdvisorAuthenticationRefreshController;
-use AdvisingApp\Ai\Http\Controllers\QnaAdvisorRequestAuthenticationController;
 use AdvisingApp\Ai\Http\Controllers\QnaAdvisors\SendAdvisorMessageController as SendQnaAdvisorMessageController;
 use AdvisingApp\Ai\Http\Controllers\QnaAdvisors\ShowAdvisorController;
 use AdvisingApp\Ai\Http\Middleware\EnsureQnaAdvisorEmbedIsEnabled;
